% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLDAShiny2.R
\name{MLDAShiny2}
\alias{MLDAShiny2}
\title{Launch 'MLDAShiny2' Interface}
\usage{
MLDAShiny2()
}
\value{
Nothing
}
\description{
MLDAShiny2() loads interactive user interface built using R 'shiny'.
}
\details{
Start MLDAShiny2

The interactive user interface is to provide an easy way for linear discriminant analysis and downloading relevant plot.
}
\examples{
if(interactive()){
library(rmarkdown)
MLDAShiny2()
}
}
\keyword{MLDAShiny2}
